Vastly refactor time
Times now contain phantom-typed fields where most of the different
behaviour (delimiters, padding, etc) is ascribed to the field position
and most of the common code is factored out into the field proper.

Needs a lot more unit testing and hooking up to the UI.

This has taken ages to get to a stable-ish state where I can move on
with zombiesplit, apologies!