Start adding next events
We now have to start a run (X), and X will 'reset' the run (or, at
least, the cursor for now).
The cursor is now no longer active when off a run.

I've blocked out M, S, and . to enter minutes, seconds, and milliseconds
but this'll require some more work.