Overhaul font rendering
The renderer now holds a lot of implicit state about where the next
font render will happen, which texture it will use, and what the font
metrics are.

Font and window metrics are now carted around in structs, in prep for
eventually leaving them to configuration and/or supporting multiple
fonts.

The metrics section has forgotten about SDL, in an attempt to decouple
some of the UI from SDL's usage.