Refactor model and add DB
Ya boi forgot to commit, and ended up with a mega-commit.

There's now an overarching 'session' type that'll serve as the full
surface contact between the presenter and the model.

There's also a database type (which does nothing but create a lot of
tables at the moment), and a clap CLI that offers a (pointless) choice
of initialising a database, adding a game to it, and starting a run.