Add list-categories function
This just retrieves a sorted rundown of all (game,category) pairs in the
database.

One day, it might return things like run counts,
last run dates, etc.  For now, though, it mainly exists
to export out information about how many categories
are in the database, and serve as the starting point of a full-run
backup system.