Move comparison sourcing to observers
We're slowly migrating everything; hopefully eventually the only use that the
split UI will have for the session is sending it control signals.

The colour used for comparison fallbacks doesn't look right yet.

Need to consolidate and remove a lot of code that's built up in the comparison
part of zombiesplit now.