Abstract event pumping out of Instance
My goal here is to make Instance turn into a generic user interface object that
can get run, and which is parametric over the exact event pump and renderer in
use, with the SDL specific code being responsible for injecting its particular
pump and renderer.

I still haven't for the life of me decided if this is MVP or MVC.  I think
the presence of UI modes fundamentally complicates things.