Reorganise footer widget
The footer widget now contains a stack of row sub-widgets, each of
which has a label and some handle to the particular part of the footer
state that contains the time.

To test this out, I've added a comparison widget, but this isn't yet
updated by anything, and so I'll need to add a hook for updating this
when the comparison is loaded.