Remove children() from Widget trait
This makes nested rendering become purely recursive, which introduces
the possibility of stack overflow but with the sorts of UI we're making
in zombiesplit that's very unlikely to happen in practice.

It does, however, simplify things in the implementation somewhat.  And
I think I overcomplicated the UI library at first pass so I'm trying to
rectify that.