Implement user-customisable time delimiters
Instead of hardcoding a delimiter after each position, we now hold the
delimiters in the same configuration as the positions, parse them as
part of the layout information, and read them off at the same time as we
read off the positions.

Massively complicates the layout parser; simplifies the rendering.
Swings and roundabouts.

This code needs some cleaning up, maybe introducing a builder struct or
something.