Start introducing tokio plumbing
This is a big commit that doesn't really do much.  The main features
include introducing serde for everything that forms the action/event
(perhaps to be renamed request/response?) interface, refactoring the UI
to try start making it easier to separate into a client, moving towards
decoupling SDL, and adding a big whack of serde boilerplate that doesn't
yet do anything but will eventually.