Drastically simplify presenter state
I actually did this quite a while back, but forgot to commit it.  Silly
me!

This is a fairly big commit that changes a lot of things, but the main
thing is the unification of presenter cursor and state cursor, backed
with removing the inactive state (all runs always start active now) and
adding some law-of-Demeter wrapping.

There are a few strange bugs that need ironing out, but we're getting
there.