Implement sum of best, refactor comparisons
Sum of best is available as a footer row, and replaces up-to-cursor
in the defaults.  It is defined as the sum of all split PBs for a
game-category.

This incorporates a fairly large refactoring of the comparison code,
pushing some things into the database and removing some unnecessary
work outside of the database that could've been done *in* the database.

PREVIOUS VERSIONS OF THE DB ARE INCOMPATIBLE.  To fix this, you will
need to create the view 'split_pb' in the database init.rs file
manually.  Maybe one day we'll have database migrations?