Add initial resizing support
The view now responds to resize requests by forcing a full re-layout
with an appropriately changed layout context.  All in all, this works
fairly well.

The root widget has been refactored to make it more amenable to
resizing.

Needs testing and fixing up, especially when it comes to splits not
all appearing on the screen or text overlapping.