Add Label pseudo-widget
This pulls in a lot of the logic for calculating the minimum bounding
box of a label, as well as rendering it given some state (in the most
straightforward case just as &str), and makes it a uniform widget.

I've only converted one label to a Label (the split name in Row), but
more will follow soon.