Add basic scrolling behaviour to the split set
This is through an intermediate iterator that produces split indices
given a number of rows and the cursor position, and tries to put the
cursor in the middle except where doing so is inefficient to do so.

This iterator isn't yet properly unit tested, nor does it yet handle
pinning of top and bottom splits (I'm not sure of the best way to do
it!).  We'll see.