Consolidate total events, implement gRPC encoding
Having two different events for sum-of-best and attempt/comparison
total seemed strange, especially when the latter two were awkwardly
fused with pace notes even though only one of them needs one.

I might strip the pace note out of the total/aggregate entirely and have
it in a separate event, eventually.

Unsure how to add these to the dump.  I think we need to dump the whole
comparison as well as the run, and have a wire format for it.