Push through most of the gRPC boilerplate
This isn't quite working yet, because we need to a) hook up the initial
state dump to the presenter; and b) finish encoding state updates.  And,
of course, c) fix the bugs, and d) clean up the spaghetti code.

But it's nearly there.  The code compiles.  zsclient can again see
zsserver.