Port to ugly
Hoo boy this has been a very long and difficult refactoring commit.
A lot of this has been teaching zombiesplit to handle colour palettes
and font maps that are hashmaps rather than structs, since ugly is
by nature a lot more general as to how it does things.

Breaking changes to configuration, which should hopefully be clear from
the code that handles them for now.  Palettes are now specified in RON
(rusty object notation), and the reason for this is slightly petty
(serde TOML can't decode enums), but we move.